fix(memento): #MA-866 custom punishments date by type.
