+++ incidents/src/main/resources/public/template/behaviours/sniplet-memento/incidents.html
@@ -64,7 +64,7 @@
                         <div class="flex-row row__15">
                             <!-- date -->
                             <div class="flex-col col__3 min-height-45">
-                                <span>[[vm.formatIncident(mementoIncident.item.date)]]</span>
+                                <span>[[vm.formatDate(mementoIncident)]]</span>
                             </div>
                             <!-- category -->
                             <div class="flex-col col__3 min-height-45">
