+++ incidents/src/main/resources/public/ts/controllers/punishments.ts
@@ -261,37 +261,7 @@ export const punishmentController = ng.controller('PunishmentController',
 
             };
 
-            vm.getPunishmentDate = (punishment: IPunishment): string => {
-                let createdDate: string = DateUtils.format(punishment.created_at, DateUtils.FORMAT['DAY-MONTH-YEAR']);
-
-                if (punishment.type) {
-                    switch (punishment.type.punishment_category_id) {
-                        case 1: // DUTY
-                            let dutyDate: string = createdDate;
-                            if ((<IPDutyField> punishment.fields).delay_at) {
-                                dutyDate = DateUtils.format((<IPDutyField> punishment.fields).delay_at, DateUtils.FORMAT['DAY-MONTH-YEAR']);
-                            }
-                            return lang.translate('incidents.punishments.date.for.the') + dutyDate;
-                        case 2: // DETENTION
-                            let startDetentionDate: string = createdDate;
-                            if ((<IPDetentionField> punishment.fields).start_at) {
-                                startDetentionDate = DateUtils.format((<IPDetentionField> punishment.fields).start_at, DateUtils.FORMAT['DAY-MONTH-YEAR']);
-                            }
-                            return lang.translate('incidents.punishments.date.for.the') + startDetentionDate;
-                        case 3: // BLAME
-                            return lang.translate('incidents.punishments.date.created.on') + createdDate;
-                        case 4: // EXCLUSION
-                            if ((<IPDetentionField> punishment.fields).start_at && (<IPDetentionField> punishment.fields).end_at) {
-                                let startExcludeDate: string = DateUtils.format((<IPExcludeField> punishment.fields).start_at, DateUtils.FORMAT['DAY-MONTH-YEAR']);
-                                let endExcludeDate: string = DateUtils.format((<IPExcludeField> punishment.fields).end_at, DateUtils.FORMAT['DAY-MONTH-YEAR']);
-                                if (startExcludeDate && endExcludeDate) {
-                                    return lang.translate('incidents.punishments.date.from') + startExcludeDate + lang.translate('incidents.punishments.date.to') + endExcludeDate;
-                                }
-                            }
-                    }
-                }
-                return lang.translate('incidents.punishments.date.created.on') + createdDate;
-            };
+            vm.getPunishmentDate = (punishment: IPunishment): string => PunishmentsUtils.getPunishmentDate(punishment);
 
             vm.getPunishmentTime = (punishment: IPunishment): string => {
                 if (punishment.type && punishment.type.punishment_category_id === 2) { // DETENTION
